+++ satella/coding/structures/dictionaries/dict_object.py
@@ -8,7 +8,7 @@ from satella.exceptions import ConfigurationValidationError
 T = tp.TypeVar('T')
 
 
-class DictObject(dict, tp.MutableMapping[str, T]):
+class DictObject(dict, tp.MutableMapping[tp.Hashable, T]):
     """
     A dictionary wrapper that can be accessed by attributes.
 
@@ -34,14 +34,14 @@ class DictObject(dict, tp.MutableMapping[str, T]):
         return 'DictObject(%s)' % (super().__repr__(),)
 
     @rethrow_as(KeyError, AttributeError)
-    def __getattr__(self, item: str) -> T:
+    def __getattr__(self, item: tp.Hashable) -> T:
         return self[item]
 
-    def __setattr__(self, key: str, value: T) -> None:
+    def __setattr__(self, key: tp.Hashable, value: T) -> None:
         self[key] = value
 
     @rethrow_as(KeyError, AttributeError)
-    def __delattr__(self, key: str) -> None:
+    def __delattr__(self, key: tp.Hashable) -> None:
         del self[key]
 
     def is_valid_schema(self, schema: tp.Optional[tp.Union[Descriptor, tp.Dict]] = None,
@@ -74,7 +74,7 @@ class DictObject(dict, tp.MutableMapping[str, T]):
             return True
 
 
-def apply_dict_object(v: tp.Union[tp.Any, tp.Dict[str, T]]) -> tp.Union[DictObject, tp.Any]:
+def apply_dict_object(v: tp.Union[tp.Any, tp.Dict[tp.Hashable, T]]) -> tp.Union[DictObject, tp.Any]:
     """
     Apply DictObject() to every dict inside v.