+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `Predicate.is_instance`
 * added `filter_out_false` and `filter_out_nones`
+* fixed typing for `DictObject`